% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallSampleEvidenceSynthesis.R
\docType{package}
\name{SmallSampleEvidenceSynthesis}
\alias{SmallSampleEvidenceSynthesis}
\alias{SmallSampleEvidenceSynthesis-package}
\title{SmallSampleEvidenceSynthesis}
\description{
SmallSampleEvidenceSynthesis
}
